Removed ranger integration.
It worked pretty well, but image previews don't for a kind of convoluted
reason.  Without the image previews, FZF and nvim-tree just work a bit
better for me inside neovim.  I still use ranger outside of nvim, and
that seems like it will be the best workflow for me.

The reason why image previews don't work in ranger running in nvim:

At the end of the day, the plugin launches ranger in neovim's embedded
terminal and relies of a library called "ueberzug" to display images.
Unfortunately, that library was discontinued by its author.

In standalone ranger, there are many options for displaying images, and
I use the kitty protocol because I use kitty as my terminal emulator.
But that won't work in the neovim terminal emulator because that
emulator doesn't support the kitty protocol.